requirements/requirements.txt
matplotlib==3.1.0
easydict==1.9
Pillow==6.1.0
imageio==2.5.0
tqdm==4.36.1
absl-py==0.8.1
tensorflow-datasets==1.2.0
ipython==7.8.0
